+++ reductstore/src/storage/query/condition/operators/logical.rs
@@ -3,10 +3,14 @@
 
 mod all_of;
 mod any_of;
+mod r#in;
+mod nin;
 mod none_of;
 mod one_of;
 
 pub(crate) use all_of::AllOf;
 pub(crate) use any_of::AnyOf;
+pub(crate) use nin::Nin;
 pub(crate) use none_of::NoneOf;
 pub(crate) use one_of::OneOf;
+pub(crate) use r#in::In;