+++ reductstore/src/storage/query/condition/parser.rs
@@ -6,7 +6,7 @@ use crate::storage::query::condition::operators::arithmetic::{
     Abs, Add, Div, DivNum, Mult, Rem, Sub,
 };
 use crate::storage::query::condition::operators::comparison::{Eq, Gt, Gte, Lt, Lte, Ne};
-use crate::storage::query::condition::operators::logical::{AllOf, AnyOf, NoneOf, OneOf};
+use crate::storage::query::condition::operators::logical::{AllOf, AnyOf, In, Nin, NoneOf, OneOf};
 use crate::storage::query::condition::operators::misc::{Cast, Exists, Ref};
 use crate::storage::query::condition::operators::string::{Contains, EndsWith, StartsWith};
 use crate::storage::query::condition::reference::Reference;
@@ -64,7 +64,15 @@ impl Parser {
                     Ok(vec![Constant::boxed(Value::String(value.clone()))])
                 }
             }
-            _ => Err(unprocessable_entity!("Invalid JSON value: {}", json)),
+
+            JsonValue::Array(_) => Err(unprocessable_entity!(
+                "Array type is not supported: {}",
+                json
+            )),
+            JsonValue::Null => Err(unprocessable_entity!(
+                "Null type is not supported: {}",
+                json
+            )),
         }
     }
 
@@ -118,6 +126,8 @@ impl Parser {
             "$none_of" => NoneOf::boxed(operands),
             "$xor" => OneOf::boxed(operands),
             "$one_of" => OneOf::boxed(operands),
+            "$in" => In::boxed(operands),
+            "$nin" => Nin::boxed(operands),
 
             // Comparison operators
             "$eq" => Eq::boxed(operands),
@@ -231,12 +241,22 @@ mod tests {
     }
 
     #[rstest]
-    fn test_parser_invalid_value(parser: Parser) {
-        let json = serde_json::from_str(r#"{"&ref": {"$and": []}}"#).unwrap();
+    fn test_parser_invalid_array_type(parser: Parser) {
+        let json = serde_json::from_str(r#"{"&label": {"$in": [10, 20]}}"#).unwrap();
         let result = parser.parse(&json);
         assert_eq!(
             result.err().unwrap().to_string(),
-            "[UnprocessableEntity] Invalid JSON value: []"
+            "[UnprocessableEntity] Array type is not supported: [10,20]"
+        );
+    }
+
+    #[rstest]
+    fn test_parser_invalid_null(parser: Parser) {
+        let json = serde_json::from_str(r#"{"&ref": {"$and": null}}"#).unwrap();
+        let result = parser.parse(&json);
+        assert_eq!(
+            result.err().unwrap().to_string(),
+            "[UnprocessableEntity] Null type is not supported: null"
         );
     }
 
@@ -260,6 +280,8 @@ mod tests {
         #[case("$none_of", "[true, true]", Value::Bool(false))]
         #[case("$xor", "[true, true]", Value::Bool(false))]
         #[case("$one_of", "[true, true]", Value::Bool(false))]
+        #[case("$in", "[\"a\", \"a\", \"b\"]", Value::Bool(true))]
+        #[case("$nin", "[\"a\", \"a\", \"b\"]", Value::Bool(false))]
         // Comparison operators
         #[case("$eq", "[10, 10]", Value::Bool(true))]
         #[case("$gt", "[20, 10]", Value::Bool(true))]